put test util methods in a separate spec_helper file
